requirements.txt
pyyaml >= 6.0
bs4 >= 0.0.1
requests >= 2.28.1
pandas >= 1.5.2
